+++ src/com/library/drawing/draw/service/DisplayService.java
@@ -1,0 +1,17 @@
+package com.library.drawing.draw.service;
+
+/**
+ * 
+ * @author Srijan Bajracharya<srijan.bajracharya@gmail.com>
+ *
+ */
+public interface DisplayService {
+
+    /**
+     * draws different figures
+     * 
+     * @param canvas
+     * @author Srijan Bajracharya<srijan.bajracharya@gmail.com>
+     */
+    public void display(char[][] canvas);
+}
