+++ src/com/library/drawing/draw/service/DrawService.java
@@ -1,40 +1,8 @@
 package com.library.drawing.draw.service;
 
 import java.awt.Canvas;
-import java.util.Map;
 
 public interface DrawService {
-
-    /**
-     * creates canvas array
-     * 
-     * @param width
-     * @param height
-     * @return {@link Canvas}
-     * @author Srijan Bajracharya<srijan.bajracharya@gmail.com>
-     */
-    public char[][] drawCanvas(int width, int height);
-
-    /**
-     * draws different figures
-     * 
-     * @param canvas
-     * @author Srijan Bajracharya<srijan.bajracharya@gmail.com>
-     */
-    public void display(char[][] canvas);
-
-    /**
-     * inserts "*" to a multidimensional array to draw rectangle
-     * 
-     * @param canvas
-     * @param x1
-     * @param y1
-     * @param x2
-     * @param y2
-     * @return {@link Canvas}
-     * @author Srijan Bajracharya<srijan.bajracharya@gmail.com>
-     */
-    public char[][] drawRectangle(char[][] canvas, int x1, int y1, int x2, int y2);
 
     /**
      * detects collision and fills
@@ -62,13 +30,5 @@ public interface DrawService {
     public char[][] removeFill(char[][] canvas, int x1, int y1, int x2, int y2);
 
     char[][] fillAll(char[][] canvas, char c);
-
-    char[][] removeLineFill(char[][] canvas, int x1, int y1, int x2, int y2, int width, int height);
-
-    Boolean isInsideRectangle(int x, int y, Map<String, Integer> coordinates);
-
-    char[][] drawLine(char[][] canvas, Map<String, Integer> coordinates);
-
-    char[][] line(char[][] canvas, Map<String, Integer> coordinates);
 
 }
